splited main into components
